[libes]
Initial commit.
