--- lib/libes/source/ETicketContentRecord_V1.cpp
@@ -1,6 +1,0 @@
-#include <es/ETicketContentRecord_V1.h>
-
-
-
-es::ETicketContentRecord_V1::ETicketContentRecord_V1()
-{}
