+++ lib/libes/source/SectionHeader_V2.cpp
@@ -1,46 +1,46 @@
-#include <es/ETicketSectionHeader_V2.h>
+#include <es/SectionHeader_V2.h>
 
 
 
-es::ETicketSectionHeader_V2::ETicketSectionHeader_V2()
+es::SectionHeader_V2::SectionHeader_V2()
 {}
 
-es::ETicketSectionHeader_V2::ETicketSectionHeader_V2(const ETicketSectionHeader_V2 & other)
+es::SectionHeader_V2::SectionHeader_V2(const SectionHeader_V2 & other)
 {
 	copyFrom(other);
 }
 
-es::ETicketSectionHeader_V2::ETicketSectionHeader_V2(const byte_t * bytes, size_t len)
+es::SectionHeader_V2::SectionHeader_V2(const byte_t * bytes, size_t len)
 {
 	importBinary(bytes, len);
 }
 
-bool es::ETicketSectionHeader_V2::operator==(const ETicketSectionHeader_V2 & other) const
+bool es::SectionHeader_V2::operator==(const SectionHeader_V2 & other) const
 {
 	return isEqual(other);
 }
 
-bool es::ETicketSectionHeader_V2::operator!=(const ETicketSectionHeader_V2 & other) const
+bool es::SectionHeader_V2::operator!=(const SectionHeader_V2 & other) const
 {
 	return !isEqual(other);
 }
 
-void es::ETicketSectionHeader_V2::operator=(const ETicketSectionHeader_V2 & other)
+void es::SectionHeader_V2::operator=(const SectionHeader_V2 & other)
 {
 	copyFrom(other);
 }
 
-const byte_t * es::ETicketSectionHeader_V2::getBytes() const
+const byte_t * es::SectionHeader_V2::getBytes() const
 {
 	return mBinaryBlob.getBytes();
 }
 
-size_t es::ETicketSectionHeader_V2::getSize() const
+size_t es::SectionHeader_V2::getSize() const
 {
 	return mBinaryBlob.getSize();
 }
 
-void es::ETicketSectionHeader_V2::exportBinary()
+void es::SectionHeader_V2::exportBinary()
 {
 	mBinaryBlob.alloc(sizeof(sSectionHeader_v2));
 	sSectionHeader_v2* hdr = (sSectionHeader_v2*)mBinaryBlob.getBytes();
@@ -52,7 +52,7 @@ void es::ETicketSectionHeader_V2::exportBinary()
 	hdr->section_type = (mSectionType);
 }
 
-void es::ETicketSectionHeader_V2::importBinary(const byte_t * bytes, size_t len)
+void es::SectionHeader_V2::importBinary(const byte_t * bytes, size_t len)
 {
 	if (len < sizeof(sSectionHeader_v2))
 	{
@@ -69,10 +69,10 @@ void es::ETicketSectionHeader_V2::importBinary(const byte_t * bytes, size_t len)
 	mRecordSize = hdr->record_size.get();
 	mSectionSize = hdr->section_size.get();
 	mRecordNum = hdr->record_num.get();
-	mSectionType = (SectionType)hdr->section_type.get();
+	mSectionType = (ticket::SectionType)hdr->section_type.get();
 }
 
-bool es::ETicketSectionHeader_V2::isEqual(const ETicketSectionHeader_V2 & other) const
+bool es::SectionHeader_V2::isEqual(const SectionHeader_V2 & other) const
 {
 	return (mSectionOffset == other.mSectionOffset) \
 		&& (mRecordSize == other.mRecordSize) \
@@ -81,7 +81,7 @@ bool es::ETicketSectionHeader_V2::isEqual(const ETicketSectionHeader_V2 & other)
 		&& (mSectionType == other.mSectionType);
 }
 
-void es::ETicketSectionHeader_V2::copyFrom(const ETicketSectionHeader_V2 & other)
+void es::SectionHeader_V2::copyFrom(const SectionHeader_V2 & other)
 {
 	if (other.getSize())
 	{
@@ -98,62 +98,62 @@ void es::ETicketSectionHeader_V2::copyFrom(const ETicketSectionHeader_V2 & other
 	}
 }
 
-void es::ETicketSectionHeader_V2::clear()
+void es::SectionHeader_V2::clear()
 {
 	mBinaryBlob.clear();
 	mSectionOffset = 0;
 	mRecordSize = 0;
 	mSectionSize = 0;
 	mRecordNum = 0;
-	mSectionType = PERMANENT;
+	mSectionType = ticket::SECTION_PERMANENT;
 }
 
-uint32_t es::ETicketSectionHeader_V2::getSectionOffset() const
+uint32_t es::SectionHeader_V2::getSectionOffset() const
 {
 	return mSectionOffset;
 }
 
-void es::ETicketSectionHeader_V2::setSectionOffset(uint32_t offset)
+void es::SectionHeader_V2::setSectionOffset(uint32_t offset)
 {
 	mSectionOffset = offset;
 }
 
-uint32_t es::ETicketSectionHeader_V2::getRecordSize() const
+uint32_t es::SectionHeader_V2::getRecordSize() const
 {
 	return mRecordSize;
 }
 
-void es::ETicketSectionHeader_V2::setRecordSize(uint32_t size)
+void es::SectionHeader_V2::setRecordSize(uint32_t size)
 {
 	mRecordSize = size;
 }
 
-uint32_t es::ETicketSectionHeader_V2::getSectionSize() const
+uint32_t es::SectionHeader_V2::getSectionSize() const
 {
 	return mSectionSize;
 }
 
-void es::ETicketSectionHeader_V2::getSectionSize(uint32_t size)
+void es::SectionHeader_V2::getSectionSize(uint32_t size)
 {
 	mSectionSize = size;
 }
 
-uint16_t es::ETicketSectionHeader_V2::getRecordNum() const
+uint16_t es::SectionHeader_V2::getRecordNum() const
 {
 	return mRecordNum;
 }
 
-void es::ETicketSectionHeader_V2::setRecordNum(uint16_t record_num)
+void es::SectionHeader_V2::setRecordNum(uint16_t record_num)
 {
 	mRecordNum = record_num;
 }
 
-es::ETicketSectionHeader_V2::SectionType es::ETicketSectionHeader_V2::getSectionType() const
+es::ticket::SectionType es::SectionHeader_V2::getSectionType() const
 {
 	return mSectionType;
 }
 
-void es::ETicketSectionHeader_V2::setSectionType(SectionType type)
+void es::SectionHeader_V2::setSectionType(ticket::SectionType type)
 {
 	mSectionType = type;
 }
